Ensure that the theme requirements are met, otherwise bail and warn the user (#69)
* Outsource functions in functions.php into separate inc/setup.php and inc/assets.php files for better organization and separation.

* Add support for replacing constant prefix to gulpfile, via convenience function getThemeConfig().

* Do not initialize the theme if its requirements are not met, and inform the user with a warning.

* Use minimum PHP version 7.0 in .travis.yml and phpcs.xml.dist.

* Use rootPath and let instead of var
