+++ .travis.yml
@@ -20,15 +20,9 @@ matrix:
   fast_finish: true
 
   include:
-    # Current $required_php_version for WordPress: 5.2.4
-    # aliased to 5.2.17
-    - php: '5.2'
-      dist: precise
-    # aliased to a recent 5.6.x version
-    - php: '5.6'
-      env: SNIFF=1
     # aliased to a recent 7.0.x version
     - php: '7.0'
+      env: SNIFF=1
     # aliased to a recent 7.2.x version
     - php: '7.2'
     # bleeding edge PHP
