+++ dev/functions.php
@@ -2,383 +2,33 @@
 /**
  * WP Rig functions and definitions
  *
+ * This file must be parseable by PHP 5.2.
+ *
  * @link https://developer.wordpress.org/themes/basics/theme-functions/
  *
  * @package wprig
  */
 
-/**
- * Sets up theme defaults and registers support for various WordPress features.
- *
- * Note that this function is hooked into the after_setup_theme hook, which
- * runs before the init hook. The init hook is too late for some features, such
- * as indicating support for post thumbnails.
- */
-function wprig_setup() {
-	/*
-	 * Make theme available for translation.
-	 * Translations can be filed in the /languages/ directory.
-	 * If you're building a theme based on wprig, use a find and replace
-	 * to change 'wprig' to the name of your theme in all the template files.
-	 */
-	load_theme_textdomain( 'wprig', get_template_directory() . '/languages' );
-
-	// Add default posts and comments RSS feed links to head.
-	add_theme_support( 'automatic-feed-links' );
-
-	/*
-	 * Let WordPress manage the document title.
-	 * By adding theme support, we declare that this theme does not use a
-	 * hard-coded <title> tag in the document head, and expect WordPress to
-	 * provide it for us.
-	 */
-	add_theme_support( 'title-tag' );
-
-	/*
-	 * Enable support for Post Thumbnails on posts and pages.
-	 *
-	 * @link https://developer.wordpress.org/themes/functionality/featured-images-post-thumbnails/
-	 */
-	add_theme_support( 'post-thumbnails' );
-
-	// This theme uses wp_nav_menu() in one location.
-	register_nav_menus(
-		array(
-			'primary' => esc_html__( 'Primary', 'wprig' ),
-		)
-	);
-
-	/*
-	 * Switch default core markup for search form, comment form, and comments
-	 * to output valid HTML5.
-	 */
-	add_theme_support(
-		'html5', array(
-			'search-form',
-			'comment-form',
-			'comment-list',
-			'gallery',
-			'caption',
-		)
-	);
-
-	// Set up the WordPress core custom background feature.
-	add_theme_support(
-		'custom-background', apply_filters(
-			'wprig_custom_background_args', array(
-				'default-color' => 'ffffff',
-				'default-image' => '',
-			)
-		)
-	);
-
-	// Add theme support for selective refresh for widgets.
-	add_theme_support( 'customize-selective-refresh-widgets' );
-
-	/*
-	 * Add support for core custom logo.
-	 *
-	 * @link https://codex.wordpress.org/Theme_Logo
-	 */
-	add_theme_support(
-		'custom-logo', array(
-			'height'      => 250,
-			'width'       => 250,
-			'flex-width'  => false,
-			'flex-height' => false,
-		)
-	);
-
-	/*
-	 * Add support for wide aligments.
-	 *
-	 * @link https://wordpress.org/gutenberg/handbook/extensibility/theme-support/
-	 */
-	add_theme_support( 'align-wide' );
-
-	/*
-	 * Add support for block color palettes.
-	 *
-	 * @link https://wordpress.org/gutenberg/handbook/extensibility/theme-support/
-	 */
-	add_theme_support( 'editor-color-palette', array(
-		array(
-			'name'  => __( 'Dusty orange', 'wprig' ),
-			'slug'  => 'dusty-orange',
-			'color' => '#ed8f5b',
-		),
-		array(
-			'name'  => __( 'Dusty red', 'wprig' ),
-			'slug'  => 'dusty-red',
-			'color' => '#e36d60',
-		),
-		array(
-			'name'  => __( 'Dusty wine', 'wprig' ),
-			'slug'  => 'dusty-wine',
-			'color' => '#9c4368',
-		),
-		array(
-			'name'  => __( 'Dark sunset', 'wprig' ),
-			'slug'  => 'dark-sunset',
-			'color' => '#33223b',
-		),
-		array(
-			'name'  => __( 'Almost black', 'wprig' ),
-			'slug'  => 'almost-black',
-			'color' => '#0a1c28',
-		),
-		array(
-			'name'  => __( 'Dusty water', 'wprig' ),
-			'slug'  => 'dusty-water',
-			'color' => '#41848f',
-		),
-		array(
-			'name'  => __( 'Dusty sky', 'wprig' ),
-			'slug'  => 'dusty-sky',
-			'color' => '#72a7a3',
-		),
-		array(
-			'name'  => __( 'Dusty daylight', 'wprig' ),
-			'slug'  => 'dusty-daylight',
-			'color' => '#97c0b7',
-		),
-		array(
-			'name'  => __( 'Dusty sun', 'wprig' ),
-			'slug'  => 'dusty-sun',
-			'color' => '#eee9d1',
-		),
-	));
-
-	/*
-	 * Optional: Disable custom colors in block color palettes.
-	 *
-	 * @link https://wordpress.org/gutenberg/handbook/extensibility/theme-support/
-	 *
-	 * add_theme_support( 'disable-custom-colors' );
-	 */
-
-	/**
-	 * Add support for font sizes.
-	 *
-	 * @link https://wordpress.org/gutenberg/handbook/extensibility/theme-support/
-	 */
-	add_theme_support( 'editor-font-sizes', array(
-		array(
-			'name'      => __( 'small', 'wprig' ),
-			'shortName' => __( 'S', 'wprig' ),
-			'size'      => 16,
-			'slug'      => 'small',
-		),
-		array(
-			'name'      => __( 'regular', 'wprig' ),
-			'shortName' => __( 'M', 'wprig' ),
-			'size'      => 20,
-			'slug'      => 'regular',
-		),
-		array(
-			'name'      => __( 'large', 'wprig' ),
-			'shortName' => __( 'L', 'wprig' ),
-			'size'      => 36,
-			'slug'      => 'large',
-		),
-		array(
-			'name'      => __( 'larger', 'wprig' ),
-			'shortName' => __( 'XL', 'wprig' ),
-			'size'      => 48,
-			'slug'      => 'larger',
-		),
-	));
-
-	/*
-	 * Optional: Add AMP support.
-	 *
-	 * Add built-in support for the AMP plugin and specific AMP features.
-	 * Control how the plugin, when activated, impacts the theme.
-	 *
-	 * @link https://wordpress.org/plugins/amp/
-	 */
-	add_theme_support( 'amp', array(
-		'comments_live_list' => true,
-	));
-
-}
-add_action( 'after_setup_theme', 'wprig_setup' );
+define( 'WPRIG_MINIMUM_WP_VERSION', '4.5' );
+define( 'WPRIG_MINIMUM_PHP_VERSION', '7.0' );
 
 /**
- * Set the embed width in pixels, based on the theme's design and stylesheet.
- *
- * @param array $dimensions An array of embed width and height values in pixels (in that order).
- * @return array
+ * Bail if requirements are not met.
  */
-function wprig_embed_dimensions( array $dimensions ) {
-	$dimensions['width'] = 720;
-	return $dimensions;
-}
-add_filter( 'embed_defaults', 'wprig_embed_dimensions' );
-
-/**
- * Returns Google Fonts used in theme.
- *
- * Has filter "wprig_google_fonts".
- *
- * @return array
- */
-function wprig_get_google_fonts() {
-
-	$fonts_default = array(
-		'Roboto Condensed' => array( '400', '400i', '700', '700i' ),
-		'Crimson Text'     => array( '400', '400i', '600', '600i' ),
-	);
-
-	/*
-	 * Filters default Google fonts.
-	 *
-	 * @param array $fonts_default array of fonts to use
-	 */
-	return apply_filters( 'wprig_google_fonts', $fonts_default );
+if ( version_compare( $GLOBALS['wp_version'], WPRIG_MINIMUM_WP_VERSION, '<' ) || version_compare( phpversion(), WPRIG_MINIMUM_PHP_VERSION, '<' ) ) {
+	require get_template_directory() . '/inc/back-compat.php';
+	return;
 }
 
 /**
- * Register Google Fonts
+ * Setup and core features configuration.
  */
-function wprig_fonts_url() {
-
-	$fonts_register = wprig_get_google_fonts();
-
-	if ( empty( $fonts_register ) ) {
-		return '';
-	}
-
-	$font_families = array();
-
-	foreach ( $fonts_register as $font_name => $font_variants ) {
-		if ( ! empty( $font_variants ) ) {
-
-			// Make sure its an array.
-			if ( ! is_array( $font_variants ) ) {
-				$font_variants = explode( ',', str_replace( ' ', '', $font_variants ) );
-			}
-
-			$font_families[] = $font_name . ':' . implode( ',', $font_variants );
-
-		} else {
-			$font_families[] = $font_name;
-		}
-	}
-
-	$query_args = array(
-		'family' => implode( '|', $font_families ),
-		'subset' => 'latin-ext',
-	);
-
-	return add_query_arg( $query_args, 'https://fonts.googleapis.com/css' );
-}
+require get_template_directory() . '/inc/setup.php';
 
 /**
- * Add preconnect for Google Fonts.
- *
- * @since Twenty Seventeen 1.0
- *
- * @param array  $urls           URLs to print for resource hints.
- * @param string $relation_type  The relation type the URLs are printed.
- * @return array $urls           URLs to print for resource hints.
+ * Theme assets management.
  */
-function wprig_resource_hints( $urls, $relation_type ) {
-	if ( wp_style_is( 'wprig-fonts', 'queue' ) && 'preconnect' === $relation_type ) {
-		$urls[] = array(
-			'href' => 'https://fonts.gstatic.com',
-			'crossorigin',
-		);
-	}
-	return $urls;
-}
-add_filter( 'wp_resource_hints', 'wprig_resource_hints', 10, 2 );
-
-/**
- * Enqueue WordPress theme styles within Gutenberg.
- */
-function wprig_gutenberg_styles() {
-
-	// Add custom fonts, used in the main stylesheet.
-	$fonts_url = wprig_fonts_url();
-	if ( ! empty( $fonts_url ) ) {
-		wp_enqueue_style( 'wprig-fonts', $fonts_url, array(), null );
-	}
-
-	// Enqueue main stylesheet.
-	wp_enqueue_style( 'wprig-base-style', get_theme_file_uri( '/css/editor-styles.css' ), array(), '20180514' );
-}
-add_action( 'enqueue_block_editor_assets', 'wprig_gutenberg_styles' );
-
-/**
- * Register widget area.
- *
- * @link https://developer.wordpress.org/themes/functionality/sidebars/#registering-a-sidebar
- */
-function wprig_widgets_init() {
-	register_sidebar( array(
-		'name'          => esc_html__( 'Sidebar', 'wprig' ),
-		'id'            => 'sidebar-1',
-		'description'   => esc_html__( 'Add widgets here.', 'wprig' ),
-		'before_widget' => '<section id="%1$s" class="widget %2$s">',
-		'after_widget'  => '</section>',
-		'before_title'  => '<h2 class="widget-title">',
-		'after_title'   => '</h2>',
-	));
-}
-add_action( 'widgets_init', 'wprig_widgets_init' );
-
-/**
- * Enqueue styles.
- */
-function wprig_styles() {
-
-	// Add custom fonts, used in the main stylesheet.
-	$fonts_url = wprig_fonts_url();
-	if ( ! empty( $fonts_url ) ) {
-		wp_enqueue_style( 'wprig-fonts', $fonts_url, array(), null );
-	}
-
-	// Enqueue main stylesheet.
-	wp_enqueue_style( 'wprig-base-style', get_stylesheet_uri(), array(), '20180514' );
-
-	// Register component styles that are printed as needed.
-	wp_register_style( 'wprig-comments', get_theme_file_uri( '/css/comments.css' ), array(), '20180514' );
-	wp_register_style( 'wprig-content', get_theme_file_uri( '/css/content.css' ), array(), '20180514' );
-	wp_register_style( 'wprig-sidebar', get_theme_file_uri( '/css/sidebar.css' ), array(), '20180514' );
-	wp_register_style( 'wprig-widgets', get_theme_file_uri( '/css/widgets.css' ), array(), '20180514' );
-	wp_register_style( 'wprig-front-page', get_theme_file_uri( '/css/front-page.css' ), array(), '20180514' );
-}
-add_action( 'wp_enqueue_scripts', 'wprig_styles' );
-
-/**
- * Enqueue scripts.
- */
-function wprig_scripts() {
-
-	// If the AMP plugin is active, return early.
-	if ( wprig_is_amp() ) {
-		return;
-	}
-
-	// Enqueue the navigation script.
-	wp_enqueue_script( 'wprig-navigation', get_theme_file_uri( '/js/navigation.js' ), array(), '20180514', false );
-	wp_script_add_data( 'wprig-navigation', 'async', true );
-	wp_localize_script( 'wprig-navigation', 'wprigScreenReaderText', array(
-		'expand'   => __( 'Expand child menu', 'wprig' ),
-		'collapse' => __( 'Collapse child menu', 'wprig' ),
-	));
-
-	// Enqueue skip-link-focus script.
-	wp_enqueue_script( 'wprig-skip-link-focus-fix', get_theme_file_uri( '/js/skip-link-focus-fix.js' ), array(), '20180514', false );
-	wp_script_add_data( 'wprig-skip-link-focus-fix', 'defer', true );
-
-	// Enqueue comment script on singular post/page views only.
-	if ( is_singular() && comments_open() && get_option( 'thread_comments' ) ) {
-		wp_enqueue_script( 'comment-reply' );
-	}
-}
-add_action( 'wp_enqueue_scripts', 'wprig_scripts' );
+require get_template_directory() . '/inc/assets.php';
 
 /**
  * Custom responsive image sizes.
