+++ gulp/php.js
@@ -2,7 +2,6 @@
 'use strict';
 
 // External dependencies
-import requireUncached from 'require-uncached';
 import pump from 'pump';
 import {src, dest} from 'gulp';
 import log from 'fancy-log';
@@ -10,12 +9,13 @@ import colors from 'ansi-colors';
 
 // Internal dependencies
 import {paths, rootPath, gulpPlugins, gulpReplaceOptions} from './constants';
+import {getThemeConfig} from './utils';
 
 // We are on the first run by default
 let isFirstRun = true;
 
 // Grab the initial config so we can check if the theme slug/name change later
-let initialConfig = require(paths.config.themeConfig);
+let initialConfig = getThemeConfig();
 
 // Stash theme config
 let themeConfig = initialConfig.theme;
@@ -26,16 +26,18 @@ let themeConfig = initialConfig.theme;
 export default function php(done) {
 
     // get a fresh copy of the config
-    const config = requireUncached(paths.config.themeConfig);
+    const config = getThemeConfig(true);
 
 	// We should rebuild if this is the first run OR the theme slug/name have changed
 	let isRebuild = isFirstRun ||
 		( themeConfig.slug !== config.theme.slug ) ||
-        ( themeConfig.name !== config.theme.name );
+        ( themeConfig.name !== config.theme.name ) ||
+        ( themeConfig.constant !== config.theme.constant );
 
 	if ( isRebuild ) {
 		themeConfig.slug = config.theme.slug;
         themeConfig.name = config.theme.name;
+        themeConfig.constant = config.theme.constant;
         log(colors.yellow(`Rebuilding ${colors.bold('all')} the PHP files, this may take a while...`));
 	} else {
         log(colors.yellow(`Rebuilding just the changed PHP files, this should be quick...`));
@@ -62,6 +64,7 @@ export default function php(done) {
         gulpPlugins.phpcs.reporter('log'),
         gulpPlugins.stringReplace('wprig', config.theme.slug, gulpReplaceOptions),
         gulpPlugins.stringReplace('WP Rig', config.theme.name, gulpReplaceOptions),
+        gulpPlugins.stringReplace('WPRIG', config.theme.constant, gulpReplaceOptions),
         dest(paths.php.dest),
     ], done);
