+++ phpcs.xml.dist
@@ -78,7 +78,7 @@
 	</rule>
 
 	<!-- Include sniffs for PHP cross-version compatibility. -->
-	<config name="testVersion" value="5.2-99.0"/>
+	<config name="testVersion" value="7.0-99.0"/>
 	<rule ref="PHPCompatibility"/>
 
 	<exclude-pattern>*/node_modules/*</exclude-pattern>
@@ -91,7 +91,7 @@
 	-->
 
 
-	<config name="testVersion" value="5.2-99.0"/>
+	<config name="testVersion" value="7.0-99.0"/>
 	<rule ref="PHPCompatibility">
 		<!-- Whitelist PHP native classes, interfaces, functions and constants which
 			 are back-filled by WP.
